# docs/api-contract.yaml
openapi: 3.1.0
info:
  title: KSA Attendance System API
  version: 0.2.0
  summary: September 2026 implementation baseline between the production frontend and backend.
  description: |
    This contract captures approved API capabilities and stable domain outcomes for
    implementation. Endpoint names may change through a documented superseding decision, but the
    authorization, state, privacy, idempotency, and source-of-truth guarantees must remain.

    The permanent QR points to a stable frontend route such as `/check-in`; it is not an
    API credential. The frontend obtains an explicit browser location and submits it to
    the backend, which makes the final attendance decision using server time.

    The production browser uses the same public origin for the UI and `/api/*`. Cookie-authenticated
    state changes also require the implementation's CSRF/origin control. Abuse-sensitive operations
    may return the documented `RATE_LIMITED` response and safe `Retry-After` under DEC-036.
servers:
  - url: /api/v1
    description: Same-origin API
...
security:
  - sessionCookie: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Match and claim an approved roster entry, register a participant, and bind this browser
      description: Queues email verification but does not require verification for attendance.
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
      responses:
        "201":
          description: Participant registered, authenticated, and device bound
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedUserEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate through the shared login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated; roles come from server state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedUserEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
  /auth/logout:
    post:
      tags: [Auth]
      summary: End the current authentication session without revoking attendance-device binding
      responses:
        "204":
          description: Logged out
        "401":
          $ref: "#/components/responses/AuthenticationError"
  /auth/session:
    get:
      tags: [Auth]
      summary: Return the current authenticated user, roles, and safe session context
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedUserEnvelope"
        "401":
          $ref: "#/components/responses/AuthenticationError"
  /auth/email-verification/resend:
    post:
      tags: [Auth]
      summary: Queue another verification email for the current unverified user
      responses:
        "202":
          description: Verification request safely accepted
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
  /auth/email-verification/confirm:
    post:
      tags: [Auth]
      summary: Confirm ownership of the email address using a single-use token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  minLength: 16
      responses:
        "204":
          description: Email verified or token already consumed for the same account
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
  /auth/password/forgot:
    post:
      tags: [Auth]
      summary: Begin password recovery without disclosing account existence unnecessarily
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier:
                  type: string
                  description: Normalized email; a reset is sent only when it belongs to a verified account
      responses:
        "202":
          description: Recovery request accepted whether or not an account is disclosed
        "429":
          $ref: "#/components/responses/RateLimitedError"
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Complete password reset; does not approve a new attendance device
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        "204":
          description: Password reset and affected sessions rotated/revoked according to policy
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
...
  /admin/sheets/retry:
    post:
      tags: [Operations]
      summary: Retry selected or failed projection jobs safely
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                jobIds:
                  type: array
                  maxItems: 100
                  items:
                    type: string
                    format: uuid
      responses:
        "202":
          description: Retry accepted
        "403":
          $ref: "#/components/responses/AuthorizationError"
  /admin/sheets/reconcile:
    post:
      tags: [Operations]
      summary: Rebuild the Sheet projection from authoritative database records
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scope, reason]
              properties:
                scope:
                  type: string
                  enum: [MASTER_REGISTER, SESSION, SUMMARY, FULL]
                sessionId:
                  type: [string, "null"]
                  format: uuid
                reason:
                  type: string
                  minLength: 3
                  maxLength: 1000
      responses:
        "202":
          description: Durable reconciliation accepted
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ksa_session
      description: Secure, HttpOnly, host-only, SameSite=Lax server session; raw value is never stored. Idle expiry is 30 days and absolute expiry is 90 days.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AuthenticationError:
      description: Authentication is missing, invalid, or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AuthorizationError:
      description: Current actor is not allowed to perform this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ConflictError:
      description: Current domain state conflicts with the requested command
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimitedError:
      description: Too many requests; retry according to safe response guidance
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Role:
      type: string
      enum: [PARTICIPANT, COURSE_REP, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    User:
      type: object
      required:
        [id, fullName, email, emailVerifiedAt, roles, status, registeredAt]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        phone:
          type: [string, "null"]
          description: Normalized and unique for participant accounts
        email:
          type: string
          format: email
        serialNumber:
          type: [string, "null"]
          pattern: "^KSA-[0-9]{2}$"
        roles:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Role"
        status:
          $ref: "#/components/schemas/UserStatus"
        emailVerifiedAt:
          type: [string, "null"]
          format: date-time
        registeredAt:
          type: string
          format: date-time
...
    RegistrationRequest:
      type: object
      required:
        [fullName, phone, email, password, serialNumber, identificationPhoto]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 150
        phone:
          type: string
          minLength: 7
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 256
        serialNumber:
          type: string
          description: Trimmed and prefix-normalized before exact KSA-XX validation; must match one eligible unclaimed roster entry
        identificationPhoto:
          type: string
          format: binary
          description: Maximum 8 MB; backend produces private JPEG quality 85 at maximum 1600 px longest edge with EXIF removed
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Normalized email or canonical KSA-XX serial number
        password:
          type: string
          format: password
...
    Error:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - STALE_STATE
            - DUPLICATE_EMAIL
            - DUPLICATE_PHONE
            - DUPLICATE_SERIAL
            - ROSTER_ENTRY_NOT_ELIGIBLE
            - ROSTER_CLAIM_CONFLICT
            - LAST_ADMIN_REQUIRED
            - RECENT_REAUTHENTICATION_REQUIRED
            - SELF_APPROVAL_FORBIDDEN
            - SESSION_ALREADY_OPEN
            - SESSION_NOT_OPEN
            - SESSION_CLOSED
            - SESSION_EXTENSION_LIMIT
            - DEVICE_CHANGE_REQUIRED
            - LOCATION_UNCERTAIN
            - LOCATION_PERMISSION_DENIED
            - LOCATION_TIMEOUT
            - LOCATION_UNAVAILABLE
            - LOCATION_UNSUPPORTED
            - CLEARLY_REMOTE
            - MANUAL_VERIFICATION_NOT_ALLOWED
            - MANUAL_VERIFICATION_EXPIRED
            - PHOTO_TOO_LARGE
            - PHOTO_UNSUPPORTED
            - RATE_LIMITED
            - SERVICE_TEMPORARILY_UNAVAILABLE
        message:
          type: string
        correlationId:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    AuthenticatedUserEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [user]
          properties:
            user:
              $ref: "#/components/schemas/User"
            attendanceDeviceStatus:
              type: [string, "null"]
              enum:
                [
                  REGISTERED_BROWSER,
                  UNRECOGNIZED_BROWSER,
                  NO_ACTIVE_DEVICE,
                  null,
                ]
...